requirements.txt
asgiref==3.3.1
boto3==1.15.18
botocore==1.18.18
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==8.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
xlrd==1.2.0